requirements-dev.txt
asgiref==3.9.1
attrs==25.4.0
certifi==2025.8.3
charset-normalizer==3.4.3
cloudinary==1.44.1
Django==5.2.6
django-cloudinary-storage==0.3.0
django-cors-headers==4.9.0
django-debug-toolbar==6.1.0
django-stubs==5.2.5
django-stubs-ext==5.2.5
django-unfold==0.65.0
djangorestframework==3.16.1
djangorestframework-stubs==3.16.3
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.10.1
gunicorn==23.0.0
idna==3.10
inflection==0.5.1
iniconfig==2.1.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
mypy==1.18.2
mypy_extensions==1.1.0
packaging==25.0
pathspec==0.12.1
pillow==11.3.0
pluggy==1.6.0
psycopg2-binary==2.9.11
Pygments==2.19.2
pytest==8.4.2
pytest-django==4.11.1
python-decouple==3.8
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.3
types-PyYAML==6.0.12.20250915
types-requests==2.32.4.20250913
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.11.0
